moves snippets from task-00 to exampleHelper file
