+++ src/components/Homework/tasks/index.js
@@ -1,3 +1,4 @@
 import properties_example from "./task-00";
 import modify_state from "./task-01";
-export default { properties_example, modify_state };
+const tasks = { properties_example, modify_state };
+export default tasks;
